refactor: simplify application page 2 css
